requirements.txt
paho-mqtt==1.6.1
requests==2.32.4
PyYAML==6.0.1
Flask==2.3.3
authlib==1.2.1
flask-session==0.5.0
flask-limiter==3.5.0
